requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
chromadb
python-dotenv
